src/requirements.txt
python-slugify
requests
urllib3